Removed test main and workbook and moved test result to test serializing data
