+++ tests/test_serializing_data.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+import unittest
+
 import datetime
 from openpyxl import Workbook
-
 
 from django_excel_tools import serializers
 
 
-# Example Usage
 class OrderExcelSerializer(serializers.ExcelSerializer):
     QR_SCANNED_CHOICES = (u'有', u'無')
     ACTIVE_CHOICES = ('Yes', 'No')
@@ -32,6 +32,9 @@ class OrderExcelSerializer(serializers.ExcelSerializer):
         max_length=5, blank=True, verbose_name='Active', choices=ACTIVE_CHOICES, case_sensitive=False
     )
 
+    def extra_clean_active(self, cleaned_value):
+        return True if cleaned_value == 'Yes' else False
+
     class Meta:
         start_index = 1
         fields = (
@@ -49,19 +52,11 @@ class OrderExcelSerializer(serializers.ExcelSerializer):
         )
 
 
-class WorkbookTesting(object):
-
-    def __init__(self):
-        self.workbook = Workbook()
-        self.worksheet = self.workbook.active
-        self._generate_title()
-        self.row_1 = ['Shop A', '170707-001-00000-0', '2017-07-07', 100, '20180101', '201801', '100', u'無', 'AB',
-                      '123/Home', 'Yes']
-        self.row_2 = ['Shop B', '170707-001-00000-1', '2017-07-08', '1000', datetime.date(2017, 1, 1), '201802', 0,
-                      u'無', None, '', 'No']
-
-    def _generate_title(self):
-        self.worksheet.append([
+class TestExpectedResult(unittest.TestCase):
+    def setUp(self):
+        workbook = Workbook()
+        worksheet = workbook.active
+        worksheet.append([
             'Shop Name',
             'Order Number',
             'Sale Date',
@@ -74,3 +69,32 @@ class WorkbookTesting(object):
             'Address',
             'Active'
         ])
+        worksheet.append([
+            'Shop A', '170707-001-00000-0', '2017-07-07', 100, '20180101',
+            '201801', '100', u'無', 'AB', '123/Home', 'Yes'
+        ])
+        worksheet.append([
+            'Shop B', '170707-001-00000-1', '2017-07-08', '1000',
+            datetime.date(2017, 1, 1), '201802', 0, u'無', None, '', 'No'
+        ])
+        worksheet.append(['', '', '', ''])
+        self.serializer = OrderExcelSerializer(worksheet)
+
+    def test_result(self):
+        self.assertEqual(self.serializer.validation_errors, [], self.serializer.validation_errors)
+        data = self.serializer.cleaned_data
+        first_row = data[0]
+        self.assertEqual(first_row['shop_name'], 'Shop A')
+        self.assertEqual(first_row['quantity'], 100)
+        self.assertEqual(first_row['default_checked'], True)
+        self.assertEqual(first_row['weight'], 100)
+        self.assertEqual(first_row['address'], '123/Home')
+        self.assertEqual(first_row['active'], True)
+
+        second_row = data[1]
+        self.assertEqual(second_row['shop_name'], 'Shop B')
+        self.assertEqual(second_row['quantity'], 1000)
+        self.assertEqual(second_row['default_checked'], False)
+        self.assertEqual(second_row['weight'], 0)
+        self.assertEqual(second_row['address'], '')
+        self.assertEqual(second_row['active'], False)
